Cleanup MSVC \W4 for core and \W3 for CLI

+++ Source/astcenc_color_quantize.cpp
@@ -68,7 +68,7 @@ static inline int quant_color_clamp(
  * @return The encoded quantized value. These are not necessarily in order; the compressor
  *         scrambles the values slightly to make hardware implementation easier.
  */
-static inline int quant_color(
+static inline uint8_t quant_color(
 	quant_method quant_level,
 	int value
 ) {
@@ -84,7 +84,7 @@ static inline int quant_color(
  * @return The encoded quantized value. These are not necessarily in order; the compressor
  *         scrambles the values slightly to make hardware implementation easier.
  */
-static inline int unquant_color(
+static inline uint8_t unquant_color(
 	quant_method quant_level,
 	int value
 ) {
